Use plain-old ncurses for the tui
We need lower-level access to do the things we need to do -- widget sets just won't cut it.

For instance, we need to only read a portion of the file into the buffer and display some part of that buffer at any one time. When the user scrolls to the bottom of the buffer, the buffer should be updated with another read from the file. This behavior would simulate that of an ordinary text editor without requiring us to load the entire file into RAM. No widget sets are going to offer this functionality

Further, we need to constantly track the cursor's position on the screen and map that to an absolute byte position in the file so that we can build instructions as we go

Building a GUI is going to be a nightmare as well.