TUI: Simulated scrolling
The TUI loads only a portion of the contents of the file into memory. As the user scrolls, we keep track of what's being displayed, and refresh the buffer when the reach the end of what we've already loaded -- in doing this, we forget about the earlier portions of that buffer which are no longer visible on screen, and load in later portions that can be scrolled to. This simulates the behavior of an ordinary text editor